refactor: dummy data 분리
