+++ apps/chat-service/src/app/(home)/meeting/page.tsx
@@ -1,83 +1,7 @@
 import Link from 'next/link';
 
 import { routes } from '@/shared/model/routes';
-
-export const MEETING_DATA = [
-    {
-        meetingUuid: 'meeting1',
-        chatUuid: 'chat1',
-        title: '아이돌 취업을 위한 모든 질문 받습니다.',
-        participants: [
-            {
-                memberUuid: 'user1',
-                name: 'User One',
-                profileUrl: 'https://picsum.photos/id/30/200/300',
-            },
-            {
-                memberUuid: 'user2',
-                name: 'User Two',
-                profileUrl: 'https://picsum.photos/id/31/200/300',
-            },
-            {
-                memberUuid: 'user3',
-                name: 'User Three',
-                profileUrl: 'https://picsum.photos/id/32/200/300',
-            },
-            {
-                memberUuid: 'user4',
-                name: 'User Four',
-                profileUrl: 'https://picsum.photos/id/33/200/300',
-            },
-            {
-                memberUuid: 'user5',
-                name: 'User Five',
-                profileUrl: 'https://picsum.photos/id/34/200/300',
-            },
-            {
-                memberUuid: 'user6',
-                name: 'User Six',
-                profileUrl: 'https://picsum.photos/id/35/200/300',
-            },
-            {
-                memberUuid: 'user7',
-                name: 'User Seven',
-                profileUrl: 'https://picsum.photos/id/36/200/300',
-            },
-            {
-                memberUuid: 'user8',
-                name: 'User Eight',
-                profileUrl: 'https://picsum.photos/id/37/200/300',
-            },
-        ],
-        createdAt: '2023-10-01T12:00:00Z',
-        updatedAt: '2023-10-01T12:00:00Z',
-        isPublic: true,
-    },
-    {
-        meetingUuid: 'meeting2',
-        chatUuid: 'chat2',
-        title: '프론트엔드 개발자 취업 Q&A',
-        participants: [
-            { memberUuid: 'user3', name: 'User Three' },
-            { memberUuid: 'user4', name: 'User Four' },
-        ],
-        createdAt: '2023-11-05T15:30:00Z',
-        updatedAt: '2023-11-05T15:30:00Z',
-        isPublic: false,
-    },
-    {
-        meetingUuid: 'meeting3',
-        chatUuid: 'chat3',
-        title: '해외 취업 준비방',
-        participants: [
-            { memberUuid: 'user5', name: 'User Five' },
-            { memberUuid: 'user6', name: 'User Six' },
-        ],
-        createdAt: '2023-12-10T09:45:00Z',
-        updatedAt: '2023-12-10T09:45:00Z',
-        isPublic: true,
-    },
-];
+import { MEETING_DATA } from '@/entities/meeting/model/constants';
 
 export default async function MeetingPage() {
     const meetingData = MEETING_DATA;
